+++ src/tool/clients/fmmlxdiagrams/DiagramActions.java
@@ -12,7 +12,14 @@ import javafx.scene.control.TextInputDialog;
 import javafx.scene.input.MouseEvent;
 import tool.clients.fmmlxdiagrams.classbrowser.ClassBrowserClient;
 import tool.clients.fmmlxdiagrams.dialogs.*;
+import tool.clients.fmmlxdiagrams.dialogs.association.AssociationDialog;
+import tool.clients.fmmlxdiagrams.dialogs.association.AssociationValueDialog;
+import tool.clients.fmmlxdiagrams.dialogs.enumeration.AddEnumerationDialog;
+import tool.clients.fmmlxdiagrams.dialogs.enumeration.DeleteEnumerationDialog;
+import tool.clients.fmmlxdiagrams.dialogs.enumeration.EditEnumerationDialog;
 import tool.clients.fmmlxdiagrams.dialogs.instancegenerator.InstanceGeneratorDialog;
+import tool.clients.fmmlxdiagrams.dialogs.operation.AddOperationDialog;
+import tool.clients.fmmlxdiagrams.dialogs.operation.ChangeBodyDialog;
 import tool.clients.fmmlxdiagrams.dialogs.results.*;
 import java.util.Optional;
 import java.util.Vector;
