+++ src/tool/clients/fmmlxdiagrams/dialogs/association/AddAssociationDialog.java
@@ -1,0 +1,1 @@
+package tool.clients.fmmlxdiagrams.dialogs.association;
