+++ src/tool/clients/fmmlxdiagrams/dialogs/association/AssociationValueDialog.java
@@ -1,4 +1,4 @@
-package tool.clients.fmmlxdiagrams.dialogs;
+package tool.clients.fmmlxdiagrams.dialogs.association;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,6 +19,7 @@ import tool.clients.fmmlxdiagrams.FmmlxAssociation;
 import tool.clients.fmmlxdiagrams.FmmlxLink;
 import tool.clients.fmmlxdiagrams.FmmlxDiagram;
 import tool.clients.fmmlxdiagrams.FmmlxObject;
+import tool.clients.fmmlxdiagrams.dialogs.CustomDialog;
 import tool.clients.fmmlxdiagrams.dialogs.results.AssociationValueDialogResult;
 import tool.clients.fmmlxdiagrams.dialogs.stringandvalue.StringValue;
