+++ src/tool/clients/fmmlxdiagrams/dialogs/operation/ChangeBodyDialog.java
@@ -1,4 +1,4 @@
-package tool.clients.fmmlxdiagrams.dialogs;
+package tool.clients.fmmlxdiagrams.dialogs.operation;
 
 import java.util.Vector;
 
@@ -16,6 +16,7 @@ import javafx.scene.control.ButtonBar.ButtonData;
 import tool.clients.fmmlxdiagrams.FmmlxDiagram;
 import tool.clients.fmmlxdiagrams.FmmlxObject;
 import tool.clients.fmmlxdiagrams.FmmlxOperation;
+import tool.clients.fmmlxdiagrams.dialogs.CustomDialog;
 import tool.clients.fmmlxdiagrams.dialogs.results.ChangeBodyDialogResult;
 import tool.clients.fmmlxdiagrams.dialogs.stringandvalue.StringValue;
